re-usability refactor of collaborator_presence_module for nokhte_session

+++ packages/backend/dart/lib/tables/_real_time_enabled/active_nokhte_sessions/queries.dart
@@ -72,7 +72,7 @@ class ActiveNokhteSessionQueries with ActiveNokhteSessionsConstants {
   }
 
   Future<List> updateCurrentPhases(
-    int newPhase, {
+    double newPhase, {
     bool shouldEditCollaboratorsInfo = false,
   }) async {
     await computeCollaboratorInformation();
